@@ -1,29 +1,24 @@
 {
   "name": "jon-template",
-  "version": "0.1.0",
+  "version": "0.5.0",
   "private": false,
-  "//dependencies": {
-    "@priolo/jon#local": "file:C://Users//iorio//Documents//Jon",
-    "@priolo/jon": "^0.2.2"
-  },
   "dependencies": {
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
-    "@mui/icons-material": "^5.4.1",
-    "@mui/material": "^5.4.1",
-    "@mui/styles": "^5.4.1",
-    "@priolo/jon": "^0.3",
-    "@priolo/jon-utils": "0.2.5",
+    "@emotion/react": "11",
+    "@emotion/styled": "11",
+    "@mui/icons-material": "5",
+    "@mui/material": "5",
+    "@priolo/jon": "^0.6.3",
+    "@priolo/jon-utils": "^0.4.1",
     "i18next": "^21.6.11",
     "i18next-browser-languagedetector": "^6.1.3",
     "i18next-http-backend": "^1.3.2",
     "js-cookie": "^2.2.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-i18next": "^11.15.4",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.3",
-    "msw": "^0.27.1"
+    "msw": "^0.27.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "^5.0.1",
+    "react-i18next": "^11.16.6",
+    "react-router-dom": "6"
   },
   "devDependencies": {
     "cross-env": "^7.0.3",
@@ -46,12 +41,17 @@
   },
   "eslintConfig": {
     "extends": [
-      "react-app"
-    ]
+      "react-app",
+      "react-app/jest"
+    ],
+    "rules": {
+      "eqeqeq": "off",
+      "react-hooks/exhaustive-deps": "off"
+    }
   },
   "engines": {
-    "node": "14",
-    "npm": "6"
+    "node": ">=14",
+    "npm": ">=6"
   },
   "browserslist": {
     "production": [
