@@ -1,10 +1,9 @@
-/* eslint eqeqeq: "off", react-hooks/exhaustive-deps: "off"*/
 import { Box, Button, TextField } from "@mui/material";
 import { Add as AddIcon } from "@mui/icons-material";
-import { rules, useValidator } from "@priolo/jon";
+
 import { useEffect } from "react";
 import { useTranslation } from "react-i18next";
-import { useHistory, useParams } from "react-router";
+import { useNavigate, useParams } from "react-router";
 
 import Form from "components/form/Form";
 import FormParagraph from "components/form/FormParagraph";
@@ -12,10 +11,11 @@ import FormRow from "components/form/FormRow";
 import UserWriterSelector from "components/selectors/UserWriterSelector";
 
 import { useConfirmationRouter } from "hooks/useConfirmationRouter";
-import { useDoc } from "stores/doc";
-import { useLayout } from "stores/layout";
-import { useRoute } from "stores/route";
 
+import docStore from "stores/doc";
+import layoutStore from "stores/layout";
+import routeStore from "stores/route";
+import { rules, useValidator, useStore } from "@priolo/jon";
 
 
 function DocDetail() {
@@ -23,11 +23,13 @@ function DocDetail() {
 	// HOOKS
 	const { id } = useParams()
 	const { t } = useTranslation()
-	const history = useHistory()
-	
-	const { state: doc, fetchById, edit, setSelectProp, canSave, isSelectChanged, save } = useDoc()
-	const { setCurrentPage } = useRoute()
-	const { setTitle } = useLayout()
+	const navigate = useNavigate()
+
+	const doc = useStore(docStore)
+	const { fetchById, edit, setSelectProp, canSave, isSelectChanged, save } = docStore
+
+	const { setCurrentPage } = routeStore
+	const { setTitle } = layoutStore
 	const titleProp = useValidator(doc.select?.title, [rules.obligatory])
 	const linkProp = useValidator(doc.select?.link, [rules.url])
 
@@ -45,19 +47,15 @@ function DocDetail() {
 
 	useConfirmationRouter(isSelectChanged)
 
-
-
 	// HANDLERS
 	const handleChangeTitle = e => setSelectProp({ name: "title", value: e.target.value })
 	const handleChangeDesc = e => setSelectProp({ name: "desc", value: e.target.value })
 	const handleChangeLink = e => setSelectProp({ name: "link", value: e.target.value })
 	const handleChangeAuthor = value => setSelectProp({ name: "author_id", value })
-	const handleClickCancel = e => history.goBack()
+	const handleClickCancel = e => navigate(-1)
 	const handleClickSave = e => save().then((success) => {
-		if (success) history.goBack()
+		if (success) navigate(-1)
 	})
-
-
 
 	// RENDER
 	if (!doc.select) return null
