@@ -1,6 +1,5 @@
 /* eslint eqeqeq: "off", react-hooks/exhaustive-deps: "off"*/
-import makeStyles from '@mui/styles/makeStyles';
-import { AppBar, Toolbar, Typography, IconButton, LinearProgress, Grid } from "@mui/material"
+import { AppBar, Toolbar, Typography, IconButton, LinearProgress, Grid, Box } from "@mui/material"
 import { Menu as MenuIcon } from "@mui/icons-material"
 
 import CentralSpace from "./CentralSpace";
@@ -8,27 +7,29 @@ import Avatar from "../app/Avatar";
 import UserHeader from "pages/user/UserHeader";
 import DocHeader from "pages/doc/DocHeader";
 
-import { useRoute } from "stores/route";
-import { useLayout } from "stores/layout";
 import LangSelector from 'components/selectors/LangSelector';
 import { useTranslation } from 'react-i18next';
 
+import layoutStore from "stores/layout";
+import routeStore from "stores/route";
+import { useStore } from "@priolo/jon";
 
 
 function Header() {
 
 	// HOOKs
-	const classes = useStyles()
 	const { t } = useTranslation()
-	const { state: layout, toggleDrawerIsOpen } = useLayout()
-	const { state: route } = useRoute()
-	
+
+	const layout = useStore(layoutStore)
+	const { toggleDrawerIsOpen } = layoutStore
+	const route = useStore(routeStore)
+
 
 	// RENDER
-	const cnAppBar = `${classes.appBar} ${layout.drawerIsOpen && layout.device == "desktop" ? classes.appBarShift : ""}`
-	
+	const sxAppBar = theme => cssAppBar(theme, layout.drawerIsOpen && layout.device == "desktop")
+
 	return (
-        <AppBar position="fixed" className={cnAppBar}>
+		<AppBar position="fixed" sx={sxAppBar}>
 			<Toolbar>
 
 				<CentralSpace
@@ -37,19 +38,19 @@ function Header() {
 						<Grid container alignItems="center" wrap="nowrap">
 
 							{!layout.drawerIsOpen && <IconButton
-                                onClick={toggleDrawerIsOpen}
-                                edge="start"
-                                className={classes.menuButton}
-                                size="large"><MenuIcon /></IconButton>}
+								onClick={toggleDrawerIsOpen}
+								edge="start"
+								sx={cssMenuButton}
+								size="large"><MenuIcon /></IconButton>}
 
-							{layout.device!="mobile" && <Typography variant="h6" noWrap className={classes.title}>
+							{layout.device != "mobile" && <Typography variant="h6" noWrap sx={cssTitle}>
 								{t(layout.title)}
 							</Typography>}
-							
+
 						</Grid>
 					}
 					renderRight={<>
-						<div className={classes.grow}></div>
+						<Box sx={cssGrow} />
 						<LangSelector />
 						<Avatar />
 					</>}
@@ -67,22 +68,21 @@ function Header() {
 			{layout.busy && <LinearProgress />}
 
 		</AppBar>
-    );
+	);
 }
 
 export default Header
 
-const useStyles = makeStyles(theme => ({
-	appBar: {
-		zIndex: theme.zIndex.drawer + 1,
-		transition: theme.transitions.create(['width', 'margin'], {
-			easing: theme.transitions.easing.sharp,
-			duration: theme.transitions.duration.leavingScreen,
-		}),
-	},
-	appBarShift: {
+
+const cssAppBar = (theme, isOpen) => ({
+	//zIndex: theme.zIndex.drawer + 1,
+	transition: theme.transitions.create(['width', 'margin'], {
+		easing: theme.transitions.easing.sharp,
+		duration: theme.transitions.duration.leavingScreen,
+	}),
+	...(isOpen && {
 		marginLeft: theme.app.drawer.width,
-		width: `calc(100% - ${theme.app.drawer.width}px)`,
+		width: `calc(100% - ${theme.app.drawer.width})`,
 		"&.mobile": {
 			marginLeft: "0px",
 			width: "100%",
@@ -91,17 +91,17 @@ const useStyles = makeStyles(theme => ({
 			easing: theme.transitions.easing.sharp,
 			duration: theme.transitions.duration.enteringScreen,
 		}),
-	},
-	menuButton: {
-		color: "inherit",
-		marginRight: "10px",
-	},
+	})
+})
 
-	title: {
-		minWidth: "150px",
-	},
-	grow: {
-		flexGrow: 1,
-	},
+const cssMenuButton = theme => ({
+	color: "inherit",
+	marginRight: "10px",
+})
 
-}));
+const cssTitle = theme => ({
+	minWidth: "150px",
+})
+const cssGrow = theme => ({
+	flexGrow: 1,
+})
