@@ -1,23 +1,25 @@
-/* eslint eqeqeq: "off", react-hooks/exhaustive-deps: "off"*/
 import { useEffect } from "react";
 import { useTranslation } from "react-i18next";
 import { Drawer, Grid, IconButton, Typography } from "@mui/material";
-import makeStyles from '@mui/styles/makeStyles';
 import { Close as CloseIcon } from '@mui/icons-material'
 
 import DocFilters from "pages/doc/DocFilters";
 
-import { useLayout } from "stores/layout"
-import { useRoute } from "stores/route";
+import layoutStore from "stores/layout";
+import routeStore from "stores/route";
+import { useStore } from "@priolo/jon";
+
 
 
 function RightDrawer() {
 
 	// HOOKs
-	const classes = useStyles()
 	const { t } = useTranslation()
-	const { state: layout, setDrawerRightIsOpen } = useLayout()
-	const { state: route, haveSearchExtra } = useRoute()
+
+	const layout = useStore(layoutStore)
+	const { setDrawerRightIsOpen } = layoutStore
+	const route = useStore(routeStore)
+	const { haveSearchExtra } = routeStore
 
 	const haveExtra = haveSearchExtra()
 
@@ -26,7 +28,7 @@ function RightDrawer() {
 		else if (haveExtra) setDrawerRightIsOpen(true)
 	}, [route.currentPage, haveExtra])
 
-	
+
 	// HANDLEs	
 	const handleClickClose = () => setDrawerRightIsOpen(false)
 
@@ -37,10 +39,10 @@ function RightDrawer() {
 	}[route.currentPage] ?? null
 
 	return (
-        <Drawer anchor="right" variant="persistent"
+		<Drawer anchor="right" variant="persistent"
 			open={layout.drawerRightIsOpen}
 		>
-			<Grid container direction="column" className={classes.container}>
+			<Grid container direction="column" sx={cssContainer}>
 
 				<Grid item container alignItems="center">
 					<Grid xs item>
@@ -55,15 +57,12 @@ function RightDrawer() {
 
 			</Grid>
 		</Drawer>
-    );
+	);
 }
 
 export default RightDrawer
 
-const useStyles = makeStyles(theme => ({
-	container: {
-		margin: `${theme.app.header.height+10}px 15px 10px 15px`,
-		width: theme.app.drawer.width,
-	}
-	
-}));
+const cssContainer = theme => ({
+	margin: `15px 15px 10px 15px`,
+	width: theme.app.drawer.width,
+})
