@@ -1,19 +1,12 @@
 /* eslint eqeqeq: "off", react-hooks/exhaustive-deps: "off"*/
 import { useEffect } from 'react';
-import {
-    FormControl,
-    Grid,
-    IconButton,
-    InputAdornment,
-    InputLabel,
-    MenuItem,
-    Select,
-} from '@mui/material';
-import makeStyles from '@mui/styles/makeStyles';
+import { FormControl, Grid, IconButton, InputAdornment, InputLabel, MenuItem, Select } from '@mui/material';
 import { Close as CloseIcon } from '@mui/icons-material'
 
-import { useUser } from 'stores/user';
 import { USER_ROLES } from "stores/user/utils"
+
+import userStore from "stores/user";
+import { useStore } from "@priolo/jon";
 
 
 function UserWriterSelector({
@@ -24,11 +17,11 @@ function UserWriterSelector({
 }) {
 
 	// HOOKs
-	const classes = useStyles()
-	const { state: user, fetchAll } = useUser()
+	const user = useStore(userStore)
+	const { fetchAll } = userStore
 
 	useEffect(() => {
-		if ( user.all.length == 0 ) fetchAll()
+		if (user.all.length == 0) fetchAll()
 	}, [])
 
 
@@ -36,7 +29,7 @@ function UserWriterSelector({
 	const handleClickClear = e => onChange("")
 	const handleChange = e => onChange(e.target.value)
 
-	
+
 	// RENDER
 	return (
 		<Grid container alignItems="center">
@@ -48,7 +41,7 @@ function UserWriterSelector({
 						value={value}
 						onChange={handleChange}
 						disabled={disabled}
-						endAdornment={<InputAdornment className={classes.selectAdornment} position="end">
+						endAdornment={<InputAdornment sx={cssSelectAdornment} position="end">
 							{value && (<IconButton size="small" onClick={handleClickClear}>
 								<CloseIcon fontSize="small" />
 							</IconButton>)}
@@ -75,12 +68,6 @@ function UserWriterSelector({
 
 export default UserWriterSelector
 
-
-
-
-
-const useStyles = makeStyles(theme => ({
-	selectAdornment: {
-		marginRight: theme.spacing(3),
-	},
-}));
+const cssSelectAdornment = theme => ({
+	marginRight: theme.spacing(3),
+})
